+++ rocket-mission-api/src/main/java/cz/muni/fi/dto/rocket/UpdateRocketDTO.java
@@ -1,8 +1,7 @@
-package cz.muni.fi.dto.create;
+package cz.muni.fi.dto.rocket;
 
-import cz.muni.fi.dto.ComponentDTO;
-import cz.muni.fi.dto.MissionDTO;
-import cz.muni.fi.dto.RocketDTO;
+import cz.muni.fi.dto.component.ComponentDTO;
+import cz.muni.fi.dto.mission.MissionDTO;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
@@ -17,7 +16,10 @@ import java.util.Set;
  * @author Tomas Bouma (469275)
  */
 
-public class CreateRocketDTO {
+public class UpdateRocketDTO {
+
+    private long id;
+
     @NotNull
     @Size(min = 3, max = 50)
     private String name;
@@ -46,6 +48,14 @@ public class CreateRocketDTO {
 
     public Set<ComponentDTO> getRequiredComponents() {
         return Collections.unmodifiableSet(requiredComponents);
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
     }
 
     public void setRequiredComponents(Set<ComponentDTO> requiredComponents) {
