+++ rocket-mission-api/src/main/java/cz/muni/fi/facade/MissionFacade.java
@@ -1,8 +1,8 @@
 package cz.muni.fi.facade;
 
-import cz.muni.fi.dto.MissionDTO;
-import cz.muni.fi.dto.create.CreateMissionDTO;
-import cz.muni.fi.dto.update.UpdateMissionDTO;
+import cz.muni.fi.dto.mission.MissionDTO;
+import cz.muni.fi.dto.mission.CreateMissionDTO;
+import cz.muni.fi.dto.mission.UpdateMissionDTO;
 import cz.muni.fi.enums.MissionProgress;
 
 import java.time.ZonedDateTime;
