+++ rocket-mission-api/src/main/java/cz/muni/fi/facade/RocketFacade.java
@@ -1,8 +1,8 @@
 package cz.muni.fi.facade;
 
-import cz.muni.fi.dto.create.CreateRocketDTO;
-import cz.muni.fi.dto.RocketDTO;
-import cz.muni.fi.dto.update.UpdateRocketDTO;
+import cz.muni.fi.dto.rocket.CreateRocketDTO;
+import cz.muni.fi.dto.rocket.RocketDTO;
+import cz.muni.fi.dto.rocket.UpdateRocketDTO;
 
 import java.util.List;
