+++ rocket-mission-service/src/test/java/cz/muni/fi/facades/ComponentFacadeImplTest.java
@@ -1,8 +1,8 @@
 package cz.muni.fi.facades;
 
-import cz.muni.fi.dto.ComponentDTO;
-import cz.muni.fi.dto.create.CreateComponentDTO;
-import cz.muni.fi.dto.update.UpdateComponentDTO;
+import cz.muni.fi.dto.component.ComponentDTO;
+import cz.muni.fi.dto.component.CreateComponentDTO;
+import cz.muni.fi.dto.component.UpdateComponentDTO;
 import cz.muni.fi.facade.ComponentFacade;
 import cz.muni.fi.services.ServiceConfiguration;
 import org.springframework.beans.factory.annotation.Autowired;
