+++ rocket-mission-service/src/test/java/cz/muni/fi/facades/MissionFacadeImplTest.java
@@ -1,12 +1,9 @@
 package cz.muni.fi.facades;
 
-import cz.muni.fi.dto.MissionDTO;
-import cz.muni.fi.dto.create.CreateMissionDTO;
-import cz.muni.fi.dto.create.CreateUserDTO;
-import cz.muni.fi.dto.update.UpdateMissionDTO;
-import cz.muni.fi.enums.LevelOfExperience;
+import cz.muni.fi.dto.mission.MissionDTO;
+import cz.muni.fi.dto.mission.CreateMissionDTO;
+import cz.muni.fi.dto.mission.UpdateMissionDTO;
 import cz.muni.fi.enums.MissionProgress;
-import cz.muni.fi.enums.Role;
 import cz.muni.fi.facade.MissionFacade;
 import cz.muni.fi.services.ServiceConfiguration;
 import org.springframework.beans.factory.annotation.Autowired;
